+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -94,7 +94,7 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-04T10:12:56.421Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-04T14:33:29.151Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")