+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQS7MFXZ1JA54BYKJ75SY1WK"
+        "01KQSPHKGDE4MZ33CGEM3NKR1J"
     )
     r"""The unique identifier of the data source"""