+++ packages/orq-rc/RELEASES.md
@@ -7529,3 +7529,13 @@ Based on:
 - [python v4.9.0-rc.20] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.20] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.20 - packages/orq-rc
+
+## 2026-05-04 14:36:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.11 (2.881.17) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.21] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.21] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.21 - packages/orq-rc